+++ docs/changes.md
@@ -12,12 +12,14 @@
 - fix for `io.import_window()` method
 - added `cut_with_cookiecutter()` to cut 2D contours.
 - fix `shapes.NormalLines()` class
+- added `vedo.
 
 
 -------------------------
 ## New/Revised Examples
 ```
 examples/basic/sliders_range.py
+examples/basic/interaction_modes.py
 examples/advanced/warp6.py
 ```